feat(database): separate files to support kanban view
