+++ packages/blocks/src/database-block/database-service.ts
@@ -9,7 +9,7 @@ import { getService } from '../__internal__/service.js';
 import { BaseService } from '../__internal__/service/index.js';
 import { asyncFocusRichText, type SerializedBlock } from '../std.js';
 import type { DatabaseBlockModel } from './database-model.js';
-import type { Cell, Column } from './types.js';
+import type { Cell, Column } from './table/types.js';
 
 export class DatabaseBlockService extends BaseService<DatabaseBlockModel> {
   initDatabaseBlock(
