+++ packages/blocks/src/database-block/index.ts
@@ -1,3 +1,3 @@
 export * from './database-block.js';
 export * from './database-model.js';
-export * from './types.js';
+export * from './table/types.js';
