+++ packages/blocks/src/database-block/table/components/cell-container.ts
@@ -2,10 +2,10 @@ import { css } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { html } from 'lit/static-html.js';
 
+import { onClickOutside } from '../../utils.js';
 import type { ColumnRendererHelper } from '../register.js';
 import { DatabaseCellElement } from '../register.js';
 import type { RowHost } from '../types.js';
-import { onClickOutside } from '../utils.js';
 
 /** affine-database-cell-container padding */
 const CELL_PADDING = 8;
