+++ packages/blocks/src/database-block/table/components/column-type/select/select-option.ts
@@ -2,11 +2,11 @@ import { css, html } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import { setupVirgoScroll } from '../../../../__internal__/utils/virgo.js';
-import { VirgoInput } from '../../../../components/virgo-input/virgo-input.js';
-import { ShadowlessElement, WithDisposable } from '../../../../std.js';
+import { setupVirgoScroll } from '../../../../../__internal__/utils/virgo.js';
+import { VirgoInput } from '../../../../../components/virgo-input/virgo-input.js';
+import { ShadowlessElement, WithDisposable } from '../../../../../std.js';
+import type { DatabaseBlockModel } from '../../../../database-model.js';
 import { SELECT_TAG_NAME_MAX_LENGTH } from '../../../consts.js';
-import type { DatabaseBlockModel } from '../../../database-model.js';
 import type { SelectTag } from '../../../types.js';
 
 @customElement('affine-database-select-option')
