+++ packages/blocks/src/database-block/table/components/edit-column-popup/edit-column-popup.ts
@@ -13,7 +13,8 @@ import { createPopper } from '@popperjs/core';
 import { html, LitElement } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
 
-import type { DatabaseBlockModel } from '../../database-model.js';
+import type { DatabaseBlockModel } from '../../../database-model.js';
+import { isDivider } from '../../../utils.js';
 import type { ColumnRendererHelper } from '../../register.js';
 import type {
   Column,
@@ -23,7 +24,6 @@ import type {
   TitleColumnAction,
 } from '../../types.js';
 import type { ColumnInsertPosition } from '../../types.js';
-import { isDivider } from '../../utils.js';
 import { ColumnTypePopup } from './column-type-popup.js';
 import { styles } from './styles.js';
 import { changeColumnType, isTitleColumn, onActionClick } from './utils.js';
