+++ packages/blocks/src/database-block/table/components/edit-column-popup/utils.ts
@@ -1,6 +1,6 @@
 import { assertExists } from '@blocksuite/store';
 
-import type { DatabaseBlockModel } from '../../database-model.js';
+import type { DatabaseBlockModel } from '../../../database-model.js';
 import type { ColumnRendererHelper } from '../../register.js';
 import type { Column, ColumnActionType, ColumnType } from '../../types.js';
 import { ColumnInsertPosition } from '../../types.js';
