+++ packages/blocks/src/database-block/table/components/row-container.ts
@@ -4,9 +4,9 @@ import { html } from 'lit';
 import { repeat } from 'lit/directives/repeat.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import { BlockElementWithService } from '../../__internal__/service/components.js';
+import { BlockElementWithService } from '../../../__internal__/service/components.js';
 import { DEFAULT_COLUMN_MIN_WIDTH } from '../consts.js';
-import type { DatabaseBlockComponent } from '../database-block.js';
+import type { DatabaseBlockComponent } from '../table-block.js';
 import { SearchState } from '../types.js';
 
 export function DataBaseRowContainer(
