+++ packages/blocks/src/database-block/table/register.ts
@@ -3,8 +3,8 @@ import type { BaseBlockModel } from '@blocksuite/store';
 import { property } from 'lit/decorators.js';
 import type { literal } from 'lit/static-html.js';
 
-import { ShadowlessElement, WithDisposable } from '../std.js';
-import type { DatabaseBlockModel } from './database-model.js';
+import { ShadowlessElement, WithDisposable } from '../../std.js';
+import type { DatabaseBlockModel } from '../database-model.js';
 import type { Cell, Column, ColumnType, RowHost } from './types.js';
 
 export abstract class DatabaseCellElement<Value> extends WithDisposable(
